+++ budgetApp1/budgetApp1/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "file (1).png",
+      "filename" : "WhatsApp Image 2024-05-30 at 19.52.24.jpg",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
